pyproject.toml

[build-system]
requires = ["setuptools>=68", "setuptools-scm", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "visionbox"
version = "0.1.0"
description = "Image classification via fine-tuning torchvision pretrained models (extracted from a notebook)"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "torch",
  "torchvision",
  "tqdm",
  "numpy",
  "pillow",
  "transformers",
  "fastapi",
  "uvicorn",
  "python-multipart",
  "ultralytics",
  "qwen-vl-utils",
  "bitsandbytes",
  "accelerate",
  "av",
]

[project.optional-dependencies]
dev = ["black", "ruff"]

[project.scripts]
visionbox-train = "visionbox.classification.train_cli:main"
visionbox-predict = "visionbox.classification.predict_cli:main"
visionbox-vqa = "visionbox.vqa.vqa_cli:main"
visionbox-caption = "visionbox.captioning.caption_cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["visionbox*"]
